// src/config/swagger-apis/types.gen.ts
export type ProductDto = {
    id?: string;
    nameEn?: string | null;
    nameAr?: string | null;
    descriptionEn?: string | null;
    descriptionAr?: string | null;
    sku?: string | null;
    barcode?: string | null;
    categoryId?: string;
    categoryNameEn?: string | null;
    categoryNameAr?: string | null;
    categoryColor?: string | null;
    brandId?: string | null;
    brandNameEn?: string | null;
    brandNameAr?: string | null;
    brandLogoUrl?: string | null;
    costPrice?: number;
    sellPrice?: number;
    stockQty?: number;
    reorderLevel?: number;
    baseUnitId?: string;
    baseUnitNameEn?: string | null;
    baseUnitNameAr?: string | null;
    baseUnitSymbol?: string | null;
    purchaseUnitId?: string | null;
    purchaseUnitNameEn?: string | null;
    purchaseUnitNameAr?: string | null;
    purchaseUnitSymbol?: string | null;
    conversionFactor?: number;
    taxId?: string;
    taxNameEn?: string | null;
    taxNameAr?: string | null;
    taxRate?: number;
    taxType?: string | null;
    active?: boolean;
    imageUrls?: Array<string> | null;
    createdAt?: string;
    updatedAt?: string | null;
};